Create project 1
read a CSV file. After reading the 
file, your program is required to complete the following tasks: 
1) Find the list of student details (ID and income) for a specific country who are in 
debt (or have debt status True) and spending more than 7 hours on any social 
media. 
2) Find the list of unique countries for a specific age group. 
3) Find the age statistics for a specific age group. The age statistics include average 
time spent in hours, standard deviation of income, and which demography (e.g. rural, sub_urban or urban) spent the lowest average time on social media for the 
specific age group. 
4) Find the platform that has the highest number of users and calculate the correlation 
between the age and the income for that user base.